+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * New functions: `inversionFromCircle` and `midCircles`.
 
+* New method for `Circle` class: `tangentsThroughExternalPoint`.
+
 
 # PlaneGeometry 1.0.0 (2020-02-01)